+++ cmd/network_validator_deposit.go
@@ -17,7 +17,7 @@ const CommandOptionDry = "dry"
 // depositCmd represents the deposit command
 var depositCmd = &cobra.Command{
 	Use:   "deposit",
-	Short: "Send Deposit transactions to activate validator",
+	Short: "Send DepositDetails transactions to activate validator",
 	Long: `After preparing wallets and deposit data, this command prepares deposit transactions to the deposit contract
 address. Remember it will need your wallet address and private keys. Thus it will deduct balance from your wallet.
 
@@ -30,8 +30,8 @@ This tool is necessary to activate new validators`,
 		}
 		nodeConf := network.MustGetNodeConfig()
 
-		valSecrets := nodeConf.GetValSecrets()
-		if valSecrets == nil || valSecrets.Deposit == nil || valSecrets.Eth1Data == nil {
+		credentials := nodeConf.ValidatorCredentials
+		if credentials == nil {
 			utils.PrintColoredError("no validator credentials exist. Did you forget to setup your validators?")
 			utils.Coloredln("    lukso network validator setup")
 			return
@@ -50,7 +50,7 @@ This tool is necessary to activate new validators`,
 			return
 		}
 
-		events, err := network.NewDepositEvents(vc.Deposit.ContractAddress, nodeConf.ApiEndpoints.ExecutionApi)
+		events, err := network.NewDepositEvents(nodeConf.DepositDetails.ContractAddress, nodeConf.ApiEndpoints.ExecutionApi)
 		if err != nil {
 			cobra.CompErrorln(fmt.Sprintf("couldn't load deposit data from contract, reason: %s", err.Error()))
 			return
@@ -58,14 +58,18 @@ This tool is necessary to activate new validators`,
 
 		fmt.Println("Past deposit events loaded", len(events.Events))
 
-		totalDeposits, err := network.Deposit(&events, valSecrets.Deposit.DepositFileLocation, valSecrets.Deposit.ContractAddress, valSecrets.Eth1Data.WalletPrivKey, nodeConf.ApiEndpoints.ExecutionApi, gasPrice, dry)
+		totalDeposits, err := network.Deposit(&events, nodeConf.DepositDetails.DepositFileLocation, nodeConf.DepositDetails.ContractAddress, nodeConf.TransactionWallet.PrivateKey, nodeConf.ApiEndpoints.ExecutionApi, gasPrice, dry)
 		if err != nil {
 			cobra.CompErrorln(err.Error())
 			return
 		}
 		if !dry {
-			fmt.Printf("You successfully deposited %d key(s)! Your keys need to be activated which takes around 8h. You can check the status by calling:\n", totalDeposits)
-			fmt.Println(utils.ConsoleInBlue("        lukso network validator describe"))
+			if totalDeposits == 0 {
+				fmt.Printf("You didn't manage to deposit any keys:\n")
+			} else {
+				fmt.Printf("You successfully deposited %d key(s)! Your keys need to be activated which takes around 8h. You can check the status by calling:\n", totalDeposits)
+				fmt.Println(utils.ConsoleInBlue("        lukso network validator describe"))
+			}
 		} else {
 			fmt.Println("THIS WAS A DRY RUN - you didn't deposit any keys")
 		}