+++ src/network/constants.go
@@ -8,10 +8,7 @@ const (
 	ChainMainNet = "mainnet"
 	ChainLocal   = "local"
 
-	CommandOptionChain              = "chain"
-	CommandOptionNodeConf           = "nodeconf"
-	CommandOptionNodeName           = "nodeName"
-	ErrMsgValidatorSecretNotPresent = "validator secret not present, did you setup your validator?"
+	ErrMsgValidatorSecretNotPresent = "Validator secret not present, did you call lukso network validator setup first?"
 	ConfigDirectory                 = "configs"
 	BootnodesDirectory              = "bootnode"
 	ConfigBranchName                = "main"