+++ src/network/deposit_event.go
@@ -5,6 +5,7 @@ import (
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/lukso-network/lukso-cli/api/gethrpc"
 	"github.com/lukso-network/lukso-cli/src/network/contracts"
+	"github.com/lukso-network/lukso-cli/src/utils"
 	"math/big"
 )
 
@@ -79,6 +80,32 @@ func (d DepositEvents) Amount(pubKey string) int64 {
 	amount := int64(0)
 	for _, e := range events {
 		amount += e.Amount.Int64()
+	}
+
+	return amount
+}
+
+func (d DepositEvents) GetUniqueKeys() []string {
+	m := make(map[string]string, 0)
+	for _, e := range d.Events {
+		m[e.Pubkey] = utils.MaybeAddHexPrefix(e.Pubkey)
+	}
+
+	l := make([]string, 0)
+
+	for _, v := range m {
+		l = append(l, v)
+	}
+	return l
+}
+
+func (d DepositEvents) TotalAmount() uint64 {
+	if len(d.Events) == 0 {
+		return 0
+	}
+	amount := uint64(0)
+	for _, e := range d.Events {
+		amount += e.Amount.Uint64()
 	}
 
 	return amount