+++ src/network/validator_setup.go
@@ -2,19 +2,15 @@ package network
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/client"
-	"github.com/docker/docker/pkg/stdcopy"
 	"github.com/manifoldco/promptui"
 	"io"
 	"os"
 	"os/exec"
 	"path"
 	"runtime"
-	"strconv"
 )
 
 // CheckAndDownloadValTool checks if validator tool is present or not. If not then download the validator tool according to platform (linux / darwin)
@@ -39,7 +35,7 @@ func GetMnemonic() (string, error) {
 	return string(output), err
 }
 
-func (valSec *ValidatorSecrets) GenerateMnemonic() error {
+func (valSec *ValidatorCredentials) GenerateMnemonic() error {
 	err := CheckAndDownloadValTool()
 	if err != nil {
 		return err
@@ -71,8 +67,7 @@ func (valSec *ValidatorSecrets) GenerateMnemonic() error {
 	return nil
 }
 
-func (valSec *ValidatorSecrets) GenerateDepositData(numberOfValidators int) error {
-
+func (valSec *ValidatorCredentials) GenerateDepositData(details *DepositDetails, numberOfValidators int) error {
 	err := CheckAndDownloadValTool()
 	if err != nil {
 		return err
@@ -80,7 +75,7 @@ func (valSec *ValidatorSecrets) GenerateDepositData(numberOfValidators int) erro
 
 	depositCmd := exec.Command("./bin/network-validator-tool", "deposit-data",
 		"--as-json-list",
-		"--fork-version", valSec.ForkVersion,
+		"--fork-version", details.ForkVersion,
 		"--source-max", fmt.Sprintf("%d", numberOfValidators),
 		"--source-min", "0",
 		"--validators-mnemonic", valSec.ValidatorMnemonic,
@@ -90,23 +85,20 @@ func (valSec *ValidatorSecrets) GenerateDepositData(numberOfValidators int) erro
 	if err != nil {
 		return err
 	}
-	err = os.WriteFile(valSec.Deposit.DepositFileLocation, commandOutput, os.ModePerm)
+	err = os.WriteFile(details.DepositFileLocation, commandOutput, os.ModePerm)
 	if err != nil {
 		return err
 	}
 	return nil
 }
 
-func (valSec *ValidatorSecrets) GenerateWallet(numberOfValidators int, password string) error {
+func (valSec *ValidatorCredentials) GenerateWallet(numberOfValidators int, password string) error {
 	err := CheckAndDownloadValTool()
 	if err != nil {
 		return err
 	}
 	nodeConfigs := MustGetNodeConfig()
-	keyStoreLocation, err := nodeConfigs.GetKeyStorePath()
-	if err != nil {
-		return err
-	}
+	keyStoreLocation := nodeConfigs.Keystore.Volume
 	walletCmd := exec.Command("./bin/network-validator-tool", "keystores",
 		"--insecure",
 		"--out-loc", keyStoreLocation,
@@ -134,7 +126,7 @@ func PullEtherealImage(ctx context.Context, client *client.Client) error {
 	return err
 }
 
-func (valSec *ValidatorSecrets) DownloadEthereal(ctx context.Context, cli *client.Client) error {
+func (valSec *ValidatorCredentials) DownloadEthereal(ctx context.Context, cli *client.Client) error {
 	reader, err := cli.ImagePull(ctx, "docker.io/wealdtech/ethereal:2.7.4", types.ImagePullOptions{})
 	if err != nil {
 		return err
@@ -142,93 +134,5 @@ func (valSec *ValidatorSecrets) DownloadEthereal(ctx context.Context, cli *clien
 
 	defer reader.Close()
 	io.Copy(os.Stdout, reader)
-	return nil
-}
-
-func (valSec *ValidatorSecrets) DoDeposit(ctx context.Context, data *DepositData, cli *client.Client, rpcEndpoint string) error {
-	depData, err := json.Marshal(data)
-	if err != nil {
-		return err
-	}
-
-	resp, err := cli.ContainerCreate(ctx, &container.Config{
-		Image: "wealdtech/ethereal:2.7.4",
-		Cmd: []string{"beacon", "deposit",
-			"--allow-unknown-contract", strconv.FormatBool(valSec.Deposit.Force),
-			"--address", valSec.Deposit.ContractAddress,
-			"--connection", rpcEndpoint,
-			"--value", valSec.Deposit.Amount,
-			"--from", valSec.Eth1Data.WalletAddress,
-			"--privatekey", valSec.Eth1Data.WalletPrivKey,
-			"--data", string(depData),
-		},
-		Tty: false,
-	}, nil, nil, nil, "")
-	if err != nil {
-		return err
-	}
-
-	if err := cli.ContainerStart(ctx, resp.ID, types.ContainerStartOptions{}); err != nil {
-		return err
-	}
-
-	statusCh, errCh := cli.ContainerWait(ctx, resp.ID, container.WaitConditionNotRunning)
-	select {
-	case err := <-errCh:
-		if err != nil {
-			return err
-		}
-	case <-statusCh:
-	}
-
-	out, err := cli.ContainerLogs(ctx, resp.ID, types.ContainerLogsOptions{ShowStdout: true, ShowStderr: true})
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println("Sending deposit for validator", data.PubKey)
-	_, err = stdcopy.StdCopy(os.Stdout, os.Stderr, out)
-	return err
-}
-
-func (valSec *ValidatorSecrets) SendDepositTxn(rpcEndpoint string) error {
-	if valSec.Eth1Data.WalletAddress == "" {
-		prompt := promptui.Prompt{
-			Label: "Enter valid wallet address",
-		}
-		eth1Address, err := prompt.Run()
-		if err != nil {
-			return err
-		}
-		valSec.Eth1Data.WalletAddress = eth1Address
-	}
-	if valSec.Eth1Data.WalletPrivKey == "" {
-		prompt := promptui.Prompt{
-			Label: "Enter valid wallet private key",
-			Mask:  '*',
-		}
-		eth1PrivateKey, err := prompt.Run()
-		if err != nil {
-			return err
-		}
-		valSec.Eth1Data.WalletPrivKey = eth1PrivateKey
-	}
-	depositData, err := ParseDepositDataFromFile(valSec.Deposit.DepositFileLocation)
-	if err != nil {
-		return err
-	}
-	dockerClient, err := GetDockerClient()
-	if err != nil {
-		return err
-	}
-	err = valSec.DownloadEthereal(context.Background(), dockerClient)
-	if err != nil {
-		return err
-	}
-	for _, data := range depositData {
-		err = valSec.DoDeposit(context.Background(), data, dockerClient, rpcEndpoint)
-		if err != nil {
-			return err
-		}
-	}
 	return nil
 }